Fix core-js issue 'delete includes in strict mode'
Fix an issue due to core-js 3.6.2 that caused the 'Array.includes'
polyfill to attempt to delete the existing implementation which is
not allowed in strict mode and causes a runtime error when its run
in IE11. Until we can resolve why this happens we should use 3.6.1
and investigate the root cause.